requirements.txt
antlr4-python3-runtime >= 4.13.1
clang >= 20.1.0
llvmlite >= 0.44.0